# docs/architecture/api-specification.yaml
openapi: 3.0.3
info:
  title: FitrahTube API
  version: 1.0.0
  description: |
    REST API for FitrahTube backend. Production implementation with 113+ endpoints across 14 controllers.

    **Key Features:**
    - Public content API for Android app (no auth required)
    - Admin/moderator content management with approval workflows
    - YouTube integration for content discovery
    - Bulk import/export functionality
    - Video validation and health monitoring
    - Comprehensive audit logging

    **Authentication:**
    - Firebase Authentication with JWT tokens
    - Role-based access: ADMIN, MODERATOR
    - Public endpoints require no authentication

    **Pagination:**
    - Cursor-based pagination for infinite scroll
    - Page-based pagination for admin interfaces

    **Caching:**
    - Categories: 1 hour TTL
    - Channels: 15 minutes TTL
    - YouTube searches: Varies by type
servers:
  - url: http://localhost:8080/api
    description: Local Development
  - url: http://192.168.1.167:8080/api
    description: Local Network (Android Testing)
  - url: https://api.albunyaan.com/api
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /admin/categories:
    get:
      summary: List all categories (hierarchical)
      operationId: listCategories
      tags: [Categories]
      responses:
        '200':
          description: Category tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Create new category
      operationId: createCategory
      tags: [Categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
...
  /admin/categories/{parentId}/subcategories:
    get:
      summary: Get subcategories of a parent
      operationId: listSubcategories
      tags: [Categories]
      parameters:
        - name: parentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Subcategories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
  /admin/categories/{id}:
    get:
      summary: Get category by ID
      operationId: getCategory
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    put:
      summary: Update category
      operationId: updateCategory
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    delete:
      summary: Delete category (fails if has subcategories)
      operationId: deleteCategory
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Category deleted

  # ============================================================================
  # CHANNELS (Admin/Moderator)
  # ============================================================================
...
  /admin/channels/category/{categoryId}:
    get:
      summary: Get channels by category
      operationId: getChannelsByCategory
      tags: [Channels]
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Channels in category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
...
  /admin/approvals/pending:
    get:
      summary: List pending approvals
      description: |
        Returns pending approvals with cursor-based pagination.

        **Cursor Behavior:**
        - Single-type queries (type=CHANNEL or type=PLAYLIST): Uses standard cursor encoding with document ID and createdAt timestamp.
        - Mixed-type queries (type omitted): Uses composite cursor that encodes separate positions for channels and playlists collections to ensure monotonic pagination across collections.

        **Ordering:**
        - Results are ordered by createdAt DESC, with document ID as secondary sort for deterministic ordering when timestamps collide.

        **Cursor Format:**
        - Cursors are opaque, URL-safe Base64 tokens. Clients should not parse or construct cursors—always use the `nextCursor` value from previous response.
      operationId: listPendingApprovals
      tags: [Approvals]
      parameters:
        - name: type
          in: query
          description: Filter by content type. Omit to get both channels and playlists merged by timestamp.
          schema:
            type: string
            enum: [CHANNEL, PLAYLIST]
        - name: category
          in: query
          description: Filter by category ID
          schema:
            type: string
        - name: limit
          in: query
          description: Number of items per page (default 20)
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          description: Opaque cursor token from previous response's pageInfo.nextCursor
          schema:
            type: string
      responses:
        '200':
          description: Pending approvals
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CursorPageDto'
                  - type: object
                    description: Cursor-paginated pending approvals
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PendingApprovalDto'
  /admin/approvals/{id}/approve:
    post:
      summary: Approve pending item
      operationId: approvePendingItem
      tags: [Approvals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reviewNotes:
                  type: string
                categoryOverride:
                  type: string
      responses:
        '200':
          description: Item approved
  /admin/approvals/{id}/reject:
    post:
      summary: Reject pending item
      operationId: rejectPendingItem
      tags: [Approvals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                reviewNotes:
                  type: string
      responses:
        '200':
          description: Item rejected

  # ============================================================================
  # PLAYER (Public)
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Authentication JWT token
  schemas:
...
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
          nullable: true
        parentCategoryId:
          type: string
          nullable: true
          description: Parent category ID (null for top-level categories)
        topLevel:
          type: boolean
          default: false
          description: True if this is a root category (no parent)
        hasSubcategories:
          type: boolean
          default: false
        icon:
          type: string
          nullable: true
          description: Emoji or icon identifier
        displayOrder:
          type: integer
          nullable: true
          description: Sort order within parent category. Null until explicitly set.
        localizedNames:
          type: object
          nullable: true
          additionalProperties:
            type: string
          description: Localized category names (en, ar, nl)
          example:
            en: Islamic Knowledge
            ar: المعرفة الإسلامية
            nl: Islamitische Kennis
        createdBy:
          type: string
          nullable: true
          description: UID of user who created this category
        updatedBy:
          type: string
          nullable: true
          description: UID of user who last updated this category
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
...
    Channel:
      type: object
      required: [id, youtubeId, name]
      properties:
        id:
          type: string
        youtubeId:
          type: string
          description: YouTube channel ID
        name:
          type: string
          description: Channel name/title
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        subscribers:
          type: integer
          format: int64
          nullable: true
          description: Subscriber count
        videoCount:
          type: integer
          nullable: true
          description: Number of videos in this channel
        categoryIds:
          type: array
          items:
            type: string
          description: List of category IDs this channel belongs to
        keywords:
          type: array
          items:
            type: string
          nullable: true
          description: Search keywords/tags for this channel. Auto-populated from YouTube tags during import if not provided.
        status:
          type: string
          enum: [APPROVED, PENDING, REJECTED]
          default: PENDING
          description: Approval status of the channel
        excludedItems:
          $ref: '#/components/schemas/ExcludedItems'
          nullable: true
          description: Items to exclude from this channel
        submittedBy:
          type: string
          nullable: true
          description: UID of user who submitted this channel
        approvedBy:
          type: string
          nullable: true
          description: UID of user who approved this channel
        approvalMetadata:
          $ref: '#/components/schemas/ApprovalMetadata'
          nullable: true
          description: Detailed approval/rejection metadata
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        displayOrder:
          type: integer
          nullable: true
          description: Integer sort order. Lower values appear first. Null until explicitly set.
...
    ExcludedItems:
      type: object
      description: Items to exclude from a channel (videos, playlists, etc.)
      properties:
        videos:
          type: array
          items:
            type: string
          description: YouTube video IDs to exclude
          default: []
        playlists:
          type: array
          items:
            type: string
          description: YouTube playlist IDs to exclude
          default: []
        liveStreams:
          type: array
          items:
            type: string
          description: YouTube livestream IDs to exclude
          default: []
        shorts:
          type: array
          items:
            type: string
          description: YouTube Shorts IDs to exclude
          default: []
        posts:
          type: array
          items:
            type: string
          description: YouTube community post IDs to exclude
          default: []

    # YouTube Search & Preview DTOs
...
    CursorPageDto:
      type: object
      description: Generic cursor-based pagination wrapper
      properties:
        data:
          type: array
          items:
            type: object
          description: Array of items (type varies by endpoint)
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    PageInfo:
      type: object
      description: Pagination metadata for cursor-based pagination
      required:
        - hasNext
      properties:
        nextCursor:
          type: string
          nullable: true
          description: Cursor for fetching next page (null if last page)
        hasNext:
          type: boolean
          description: Whether more pages are available
        totalCount:
          type: integer
          nullable: true
          description: Total count of items matching current filters (optional, returned by some endpoints)
        truncated:
          type: boolean
          nullable: true
          description: |
            Indicates whether results were truncated due to safety limits.
            When true, the totalCount and results may be incomplete.
            Used by workspace exclusions aggregation to signal when hard limits
            (e.g., 500 channels, 1000 playlists) have been hit.
...
    ApprovalMetadata:
      type: object
      description: Metadata about approval/rejection actions (embedded in Channel/Playlist documents)
      properties:
        reviewedBy:
          type: string
          nullable: true
          description: Firebase UID of the reviewer
        reviewerDisplayName:
          type: string
          nullable: true
          description: Display name of the reviewer (cached)
        reviewedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the review happened
        reviewNotes:
          type: string
          nullable: true
          description: Review notes from the moderator
        rejectionReason:
          type: string
          nullable: true
          enum: [NOT_ISLAMIC, LOW_QUALITY, DUPLICATE, OTHER]
          description: Rejection reason (for rejected items)
    PendingApprovalDto:
      type: object
      description: Pending approval item
      properties:
        id:
          type: string
        type:
          type: string
          enum: [CHANNEL, PLAYLIST, VIDEO]
        entityId:
          type: string
          description: ID of the actual channel/playlist/video
        title:
          type: string
        category:
          type: string
        submittedAt:
          type: string
          format: date-time
        submittedBy:
          type: string
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata (subscribers, video count, etc.)
...